Refactoring to extract the model parts into the modules in the model folder
