+++ war/src/main/js/widgets/config/model/util.js
@@ -1,0 +1,5 @@
+
+exports.toId = function(string) {
+    string = string.trim();
+    return 'config_' + string.replace(/[\W_]+/g, '_').toLowerCase();
+};
